requirements.txt
openai-whisper
ffmpeg-python
torch
torchaudio
